refactor(FE): `useFakemonsByUser` hook
