+++ frontend/src/components/Gyms/Gyms.js
@@ -3,8 +3,9 @@ import Button from "react-bootstrap/Button";
 import Card from "react-bootstrap/Card";
 import { Link } from "react-router-dom";
 import { useCreateGym } from "../../hooks/mutations";
-import { useFakemonsByUser, useIsRegistered } from "../../hooks/queries";
+import { useIsRegistered } from "../../hooks/queries";
 import useAuth from "../../hooks/useAuth";
+import useFakemonsByUser from "../../hooks/useFakemonsByUser";
 import useGyms from "../../hooks/useGyms";
 import Alert from "../Alert";
 import ButtonWithLoader from "../ButtonWithLoader";
